fix(web): fix export blockquote spacing and async diagram copy
Restore vertical spacing between alert blockquotes in exported HTML by
only resetting horizontal blockquote margins in base.css.

Wait for mermaid/plantuml/infographic/math rendering before WeChat
clipboard copy and HTML/PDF export; strip unresolved loading placeholders.
